test(cloud): fix workers-types TS errors in STT route test; cover validation gates + ElevenLabs lane
The cloud-api tsconfig merges workers-types/DOM/bun globals, which made
`value instanceof File` on FormDataEntryValue a TS2358 and left
Response#json()'s generic inferring `undefined` at bare expect() sites.
Replace the instanceof with a structural isFilePart guard and pin json
reads to `unknown` via readJson — no `any` casts, no ts-ignore.

Rename route.whisper.test.ts -> route.test.ts: the suite now drives the
whole real route handler — shared upload-validation gates (multipart,
25MB cap, declared-MIME, magic-number + spoofed-extension rejection),
auth-failure mapping, the free whisper lane (#14806 timestamps + J3
malformed-200 boundary), and the billed ElevenLabs lane (billing wiring,
402 insufficient credits, 429/402/503/500 error mapping with reservation
refunds), lifting route.ts changed-file coverage from 38% to 90%.
